requirements.txt
aniso8601==4.0.1
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
py-postgresql==1.2.1
pytz==2018.7
six==1.12.0
Werkzeug==0.14.1
